# Conflicts:
#	modules/Northwind/MeshWeaver.Northwind.ViewModel/NorthwindDashboardArea.cs
#	modules/Northwind/MeshWeaver.Northwind.ViewModel/NorthwindDocumentationConfiguration.cs
#	modules/Northwind/MeshWeaver.Northwind.ViewModel/NorthwindViewModels.cs
#	src/MeshWeaver.Blazor/MeshWeaver.Blazor.csproj
#	src/MeshWeaver.Blazor/NavGroup.razor
#	src/MeshWeaver.Layout/Controls.cs
#	src/MeshWeaver.Layout/ExpandableControl.cs
#	src/MeshWeaver.Layout/NavMenuControl.cs

+++ modules/Northwind/MeshWeaver.Northwind.Domain/Product.cs
@@ -30,7 +30,7 @@ namespace MeshWeaver.Northwind.Domain
         [property:DisplayName("Supplier")][property: Dimension(typeof(Supplier))] int SupplierId,
         [property: DisplayName("Category")][property: Dimension(typeof(Category))] int CategoryId,
         string QuantityPerUnit,
-        decimal UnitPrice,
+        double UnitPrice,
         short UnitsInStock,
         short UnitsOnOrder,
         short ReorderLevel,
